fix suggestions from code review
